requirements.txt
streamlit
google-generativeai
pandas
python-docx
pdfplumber
openpyxl